infrastructure/ai-coach/requirements.txt
aws-cdk-lib==2.215.0
constructs>=10.0.0,<11.0.0
boto3>=1.28.0
aws-cdk.lambda-layer-kubectl-v32>=2.0.0
